[issue-244-refactor-network-handling-code] fixed broken compilation, long -> bool cast on acceptBlockHeader() case return
