+++ src/test/DoS_tests.cpp
@@ -22,6 +22,7 @@
 #include <boost/date_time/posix_time/posix_time_types.hpp>
 #include <boost/foreach.hpp>
 #include <boost/test/unit_test.hpp>
+#include <messages/NetworkMessage.h>
 
 CService ip(uint32_t i)
 {
@@ -61,7 +62,7 @@ BOOST_AUTO_TEST_CASE(DoS_banning)
     CAddress addr1(ip(0xa0b0c001));
     CNode dummyNode1(INVALID_SOCKET, addr1, "", true);
     dummyNode1.nVersion = 1;
-    Misbehaving(dummyNode1.GetId(), 100); // Should get banned
+    Network::misbehaving(dummyNode1.GetId(), 100); // Should get banned
     SendMessages(&dummyNode1);
     BOOST_CHECK(CNode::IsBanned(addr1));
     BOOST_CHECK(!CNode::IsBanned(ip(0xa0b0c001|0x0000ff00))); // Different IP, not banned
@@ -69,11 +70,11 @@ BOOST_AUTO_TEST_CASE(DoS_banning)
     CAddress addr2(ip(0xa0b0c002));
     CNode dummyNode2(INVALID_SOCKET, addr2, "", true);
     dummyNode2.nVersion = 1;
-    Misbehaving(dummyNode2.GetId(), 50);
+    Network::misbehaving(dummyNode2.GetId(), 50);
     SendMessages(&dummyNode2);
     BOOST_CHECK(!CNode::IsBanned(addr2)); // 2 not banned yet...
     BOOST_CHECK(CNode::IsBanned(addr1));  // ... but 1 still should be
-    Misbehaving(dummyNode2.GetId(), 50);
+    Network::misbehaving(dummyNode2.GetId(), 50);
     SendMessages(&dummyNode2);
     BOOST_CHECK(CNode::IsBanned(addr2));
 }
@@ -85,13 +86,13 @@ BOOST_AUTO_TEST_CASE(DoS_banscore)
     CAddress addr1(ip(0xa0b0c001));
     CNode dummyNode1(INVALID_SOCKET, addr1, "", true);
     dummyNode1.nVersion = 1;
-    Misbehaving(dummyNode1.GetId(), 100);
+    Network::misbehaving(dummyNode1.GetId(), 100);
     SendMessages(&dummyNode1);
     BOOST_CHECK(!CNode::IsBanned(addr1));
-    Misbehaving(dummyNode1.GetId(), 10);
+    Network::misbehaving(dummyNode1.GetId(), 10);
     SendMessages(&dummyNode1);
     BOOST_CHECK(!CNode::IsBanned(addr1));
-    Misbehaving(dummyNode1.GetId(), 1);
+    Network::misbehaving(dummyNode1.GetId(), 1);
     SendMessages(&dummyNode1);
     BOOST_CHECK(CNode::IsBanned(addr1));
     mapArgs.erase("-banscore");
@@ -107,7 +108,7 @@ BOOST_AUTO_TEST_CASE(DoS_bantime)
     CNode dummyNode(INVALID_SOCKET, addr, "", true);
     dummyNode.nVersion = 1;
 
-    Misbehaving(dummyNode.GetId(), 100);
+    Network::misbehaving(dummyNode.GetId(), 100);
     SendMessages(&dummyNode);
     BOOST_CHECK(CNode::IsBanned(addr));
